Clean up templates
Use inheritance, and don't set title from python. Also, since we want
the timestamp everywhere let's make it a default context processor, and
also disregard the intermediary "now" variable.

I changed it to an "isoformat()" instead. It was pretty close, and it
gets the job done for debug stuff like this.